bitvmx_wallet::wallet::wallet
Trait RegtestWallet

pub trait RegtestWallet {
    // Required methods
    fn check_network(&self) -> Result<(), WalletError>;
    fn export_wallet(&self) -> Result<(Vec<String>, Vec<String>), WalletError>;
    fn mine_to_address(
        &self,
        num_blocks: u64,
        address: &str,
    ) -> Result<Vec<Transaction>, WalletError>;
    fn mine(&self, num_blocks: u64) -> Result<Vec<Transaction>, WalletError>;
    fn fund(&mut self) -> Result<(), WalletError>;
    fn fund_address(
        &mut self,
        to_address: &str,
        amount: u64,
    ) -> Result<Transaction, WalletError>;
    fn fund_destination(
        &mut self,
        destination: Destination,
    ) -> Result<Transaction, WalletError>;
    fn clear_db(wallet_config: &WalletConfig) -> Result<(), WalletError>;
}

Extension trait for regtest-specific wallet functions This trait provides utilities that are only available in regtest mode

Required Methods

fn check_network(&self) -> Result<(), WalletError>

Check if the wallet is in regtest mode

fn export_wallet(&self) -> Result<(Vec<String>, Vec<String>), WalletError>

Export the wallet

fn mine_to_address( &self, num_blocks: u64, address: &str, ) -> Result<Vec<Transaction>, WalletError>

Generate blocks and send the coinbase reward to a specific address Returns the coinbase transactions This function is only available in regtest mode

fn mine(&self, num_blocks: u64) -> Result<Vec<Transaction>, WalletError>

Generate a specified number of blocks to a default address Returns the coinbase transactions This function is only available in regtest mode

fn fund(&mut self) -> Result<(), WalletError>

Fund the wallet with 150 BTC This function is only available in regtest mode

fn fund_address( &mut self, to_address: &str, amount: u64, ) -> Result<Transaction, WalletError>

👎Deprecated since 0.2.0: Use fund_destination instead

Send funds to a specific address and mines 1 block This function is only available in regtest mode

fn fund_destination( &mut self, destination: Destination, ) -> Result<Transaction, WalletError>

Send funds to a specific destination and mines 1 block This function is only available in regtest mode

fn clear_db(wallet_config: &WalletConfig) -> Result<(), WalletError>

Clear the database This function is only available in regtest mode

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.